requirements.txt
pytesseract
opencv-python
pillow
streamlit==1.20.0
scikit-learn==1.2.2
numpy==1.24.3
pandas==1.5.1
altair<5
